+++ src/contract/polls.cpp
@@ -4,7 +4,8 @@
 #include "main.h"
 #include "polls.h"
 #include "contract.h"
-#include "bitcoinrpc.h"
+#include "rpcclient.h"
+#include "rpcserver.h"
 #include "appcache.h"
 #include "cpid.h"
 #include "init.h" // for pwalletMain
