+++ src/main.cpp
@@ -16,7 +16,8 @@
 #include "global_objects_noui.hpp"
 #include "util.h"
 #include "cpid.h"
-#include "bitcoinrpc.h"
+#include "rpcserver.h"
+#include "rpcclient.h"
 #include "json/json_spirit_value.h"
 #include "boinc.h"
 #include "beacon.h"
