+++ src/qt/rpcconsole.cpp
@@ -3,7 +3,9 @@
 
 #ifndef Q_MOC_RUN
 #include "clientmodel.h"
-#include "bitcoinrpc.h"
+#include "rpcserver.h"
+#include "rpcclient.h"
+#include "rpcprotocol.h"
 #include "guiutil.h"
 #include "main.h"
 #endif
