+++ src/rpcdataacq.cpp
@@ -5,7 +5,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "main.h"
-#include "bitcoinrpc.h"
+#include "rpcserver.h"
 #include "cpid.h"
 #include "kernel.h"
 #include "block.h"
