+++ src/rpcrawtransaction.cpp
@@ -6,7 +6,8 @@
 #include <boost/assign/list_of.hpp>
 
 #include "base58.h"
-#include "bitcoinrpc.h"
+#include "rpcserver.h"
+#include "rpcprotocol.h"
 #include "txdb.h"
 #include "init.h"
 #include "main.h"
