+++ src/rpcwallet.cpp
@@ -7,7 +7,8 @@
 #include "txdb.h"
 #include "wallet.h"
 #include "walletdb.h"
-#include "bitcoinrpc.h"
+#include "rpcserver.h"
+#include "rpcprotocol.h"
 #include "init.h"
 #include "base58.h"
 #include "util.h"
