+++ src/wallet.cpp
@@ -15,7 +15,9 @@
 #include <boost/thread.hpp>
 #include "cpid.h"
 #include "block.h"
-#include "bitcoinrpc.h"
+#include "rpcserver.h"
+#include "rpcclient.h"
+#include "rpcprotocol.h"
 #include "util.h"
 #include "grcrestarter.h"
 #include <boost/variant/apply_visitor.hpp>
@@ -569,14 +571,14 @@ bool CWallet::AddToWallet(const CWalletTx& wtxIn)
         std::string strCmd = GetArg("-walletnotify", "");
         // Reward Sharing - Added 08-21-2016
         // if (IsCoinBase() || IsCoinStake())
-       
+
         if (wtxIn.IsCoinBase() || wtxIn.IsCoinStake())
         {
             if (fDebug10) LogPrintf("\nCoinBase:CoinStake\n");
             CBlockIndex* pBlk = mapBlockIndex[wtxIn.hashBlock];
             CBlock blk;
             bool r = blk.ReadFromDisk(pBlk);
-            if (r) 
+            if (r)
             {
                 MiningCPID bb = DeserializeBoincBlock(blk.vtx[0].hashBoinc,blk.nVersion);
                 double dResearch = bb.ResearchSubsidy + bb.InterestSubsidy;
@@ -762,7 +764,7 @@ CTxDestination GetCoinstakeDestination(const CWalletTx* wtx,CTxDB& txdb)
             {
                 if (prevout.n < prev.vout.size())
                 {
-                    //Inputs: 
+                    //Inputs:
                     const CTxOut &vout = prev.vout[prevout.n];
                     CTxDestination address;
                     if (ExtractDestination(vout.scriptPubKey, address))
@@ -815,7 +817,7 @@ void CWalletTx::GetAmounts2(list<COutputEntry>& listReceived,
         CTxDestination address;
         if (IsCoinStake())
         {
-            // R Halford - For CoinStake we must extract the address from the input 
+            // R Halford - For CoinStake we must extract the address from the input
             address = GetCoinstakeDestination(this,txdb);
         }
         else
@@ -838,13 +840,13 @@ void CWalletTx::GetAmounts2(list<COutputEntry>& listReceived,
 
         // If we are receiving the output, add it as a "received" entry
         if (fIsMine || IsCoinStake())
-        {   
+        {
             if (IsCoinStake())
-            {   
+            {
                 // For CoinStake, we must calculate the subsidy based on Net Earned due to splitstakes and empty stakes
                 output.amount += -nFee;
                 nFee=0;
-                if (output.amount != 0) 
+                if (output.amount != 0)
                 {
                         listReceived.push_back(output);
                         break;
@@ -1109,7 +1111,7 @@ void CWallet::ReacceptWalletTransactions()
         }
     }
 }
-    
+
 
 void CWalletTx::RelayWalletTransaction(CTxDB& txdb)
 {
@@ -1254,7 +1256,7 @@ void CWallet::AvailableCoins(vector<COutput>& vCoins, bool fOnlyConfirmed, const
         {
             const CWalletTx* pcoin = &(*it).second;
 			int nDepth = pcoin->GetDepthInMainChain();
-		
+
 			if (!fIncludeStakedCoins)
 			{
 				if (!IsFinalTx(*pcoin))
@@ -1280,7 +1282,7 @@ void CWallet::AvailableCoins(vector<COutput>& vCoins, bool fOnlyConfirmed, const
             for (unsigned int i = 0; i < pcoin->vout.size(); i++)
 			{
                 if ((!(pcoin->IsSpent(i)) && IsMine(pcoin->vout[i]) && pcoin->vout[i].nValue >= nMinimumInputValue &&
-                   (!coinControl || !coinControl->HasSelected() || coinControl->IsSelected((*it).first, i))) 
+                   (!coinControl || !coinControl->HasSelected() || coinControl->IsSelected((*it).first, i)))
 	     	 	   || (fIncludeStakedCoins && pcoin->IsCoinStake() && pcoin->GetBlocksToMaturity() > 0 && pcoin->GetDepthInMainChain() > 0))
 				   {
 				        vCoins.push_back(COutput(pcoin, i, nDepth));
@@ -1520,7 +1522,7 @@ bool CWallet::SelectCoins(int64_t nTargetValue, unsigned int nSpendTime, set<pai
 }
 
 // Select some coins without random shuffle or best subset approximation
-bool CWallet::SelectCoinsForStaking(int64_t nTargetValueIn, unsigned int nSpendTime, 
+bool CWallet::SelectCoinsForStaking(int64_t nTargetValueIn, unsigned int nSpendTime,
     std::set<pair<const CWalletTx*,unsigned int> >& setCoinsRet, int64_t& nValueRet) const
 {
     vector<COutput> vCoins;
@@ -1528,14 +1530,14 @@ bool CWallet::SelectCoinsForStaking(int64_t nTargetValueIn, unsigned int nSpendT
 
     setCoinsRet.clear();
     nValueRet = 0;
-    
+
     int64_t nTargetValue = nTargetValueIn;
 
-    //if (GlobalCPUMiningCPID.cpid != "INVESTOR"  && msMiningErrors7 != "Probing coin age") 
+    //if (GlobalCPUMiningCPID.cpid != "INVESTOR"  && msMiningErrors7 != "Probing coin age")
     //{
     //      nTargetValue = nTargetValueIn/2;
     //}
-    
+
     for(const COutput& output : vCoins)
     {
         const CWalletTx *pcoin = output.tx;
@@ -1567,7 +1569,7 @@ bool CWallet::SelectCoinsForStaking(int64_t nTargetValueIn, unsigned int nSpendT
     return true;
 }
 
-bool CWallet::CreateTransaction(const vector<pair<CScript, int64_t> >& vecSend, CWalletTx& wtxNew, CReserveKey& reservekey, 
+bool CWallet::CreateTransaction(const vector<pair<CScript, int64_t> >& vecSend, CWalletTx& wtxNew, CReserveKey& reservekey,
     int64_t& nFeeRet, const CCoinControl* coinControl)
 {
 
@@ -1751,14 +1753,14 @@ bool CWallet::GetStakeWeight(uint64_t& nWeight)
     //Retrieve CPID RSA_WEIGHT
     int64_t RSA_WEIGHT = IsV8Enabled(nBestHeight+1) ? 0 : GetRSAWeightByCPID(GlobalCPUMiningCPID.cpid);
     ////////////////////////////////////////////////////////////////////////////////
-    
+
     LOCK2(cs_main, cs_wallet);
     for (auto const& pcoin : setCoins)
     {
         CTxIndex txindex;
         if (!txdb.ReadTxIndex(pcoin.first->GetHash(), txindex))
             continue;
-        //1-13-2015 
+        //1-13-2015
         if (IsProtocolV2(nBestHeight+1))
         {
             if (nCurrentTime - pcoin.first->nTime > nStakeMinAge)
@@ -1778,8 +1780,8 @@ bool CWallet::GetStakeWeight(uint64_t& nWeight)
             }
         }
     }
-    
-    
+
+
     return true;
 }
 
@@ -1794,7 +1796,7 @@ void NetworkTimer()
     mdMachineTimerLast = GetAdjustedTime();
     if (elapsed < 1) elapsed = 1;
     mdPORNonce += (elapsed*10);
-    if (mdPORNonce > 2147483000) 
+    if (mdPORNonce > 2147483000)
     {
             LogPrintf("Resetting...");
             mdPORNonce=0;
@@ -1807,8 +1809,7 @@ bool CWallet::CommitTransaction(CWalletTx& wtxNew, CReserveKey& reservekey)
 {
     if(fDevbuildCripple)
     {
-        error("CommitTransaction(): Development build restrictions in effect");
-        return false;
+        return error("CommitTransaction(): Development build restrictions in effect");
     }
     {
         LOCK2(cs_main, cs_wallet);
@@ -2529,7 +2530,7 @@ std::vector<std::pair<CBitcoinAddress, CBitcoinSecret>> CWallet::GetAllPrivateKe
         assert(keypool.vchPubKey.IsValid());
         CKeyID keyID = keypool.vchPubKey.GetID();
 
-        if (!HaveKey(keyID))  
+        if (!HaveKey(keyID))
         {
             LogPrintf("GetAllPrivateKeys: Unknown key in key pool\n");
         }
